requirements.txt
python<3.13
Django>=4.2
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
psycopg2-binary>=2.9.9  
gunicorn>=21.2.0  
whitenoise>=6.6.0  
django-cors-headers>=4.3.1 
drf-yasg>=1.21.7  
django-environ>=0.12.0
requests>2.2.0
uuid=1.30
black>=25.1.0
dj-database-url=2.3.0
